x86: allow to set virtual disk size
When creating virtual disk images, we make the virtual disk just large
enough to hold both kernel and root partition.

If we are to increase a partition's size, it will make the sysupgrade
fail because there's simply not enough space in the virtual disk to
hold the new partiton layout.

Therefore, introduce new config options VDI_IMAGES_VIRTUAL_SIZE and
VMDK_IMAGES_VIRTUAL_SIZE. If set, the respective virtual disk image
will be resized to the given size. As it is a virtual disk, the image
size will not really increase.